fix: move svg files to icons folder & resolve TS error by extending row type for sorting
