+++ src/shared/ui/icons/ChevronRightIcon.tsx
