+++ src/shared/ui/icons/ClearIcon.tsx
