+++ src/shared/ui/icons/CloseIcon.tsx
