+++ src/shared/ui/icons/SearchIcon.tsx
