+++ src/shared/ui/icons/index.ts
@@ -1,0 +1,7 @@
+export { ChevronDownIcon } from './ChevronDownIcon';
+export { ChevronUpIcon } from './ChevronUpIcon';
+export { ChevronLeftIcon } from './ChevronLeftIcon';
+export { ChevronRightIcon } from './ChevronRightIcon';
+export { SearchIcon } from './SearchIcon';
+export { CloseIcon } from './CloseIcon';
+export { ClearIcon } from './ClearIcon';
